Added quit screen to main menu exit and pause screen return to menu
