+++ Assets/Scripts/UI/PauseScreen.cs
@@ -1,7 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.SceneManagement;
 
 public class PauseScreen : MonoBehaviour
 {
@@ -13,26 +12,5 @@ public class PauseScreen : MonoBehaviour
   public void Restart()
   {
     GameStateController.HandleTrigger(StateTrigger.Reset);
-  }
-
-  public void MainMenu()
-  {
-    StartCoroutine(LoadYourAsyncScene("MainMenu"));
-  }
-
-  /// <summary>Loads the scene associated with the string asycronously.</summary>
-  /// <param name="scene">The name of the scene to load.</param>
-  IEnumerator LoadYourAsyncScene(string scene)
-  {
-    // The Application loads the Scene in the background as the current Scene runs.
-    // This is particularly good for creating loading screens.
-    // You could also load the Scene by using sceneBuildIndex. In this case Scene2 has
-    // a sceneBuildIndex of 1 as shown in Build Settings.
-    AsyncOperation asyncLoad = SceneManager.LoadSceneAsync(scene);
-    // Wait until the asynchronous scene fully loads
-    while (!asyncLoad.isDone)
-    {
-      yield return null;
-    }
   }
 }
