+++ Assets/Scripts/UI/QuitScreen.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 64d4de3cb672be24e8d34f9091336d0e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
